# packages/contracts/openapi/voice-v1.json
{
  "components": {
    "schemas": {
      "VoiceContext": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "additionalProperties": false,
        "properties": {
          "agentEpoch": {
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "type": "integer"
          },
          "agentParticipantId": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "callerParticipantId": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "callId": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "language": {
            "maxLength": 32,
            "minLength": 1,
            "type": "string"
          },
          "permittedTools": {
            "items": {
              "const": "create_business_request",
              "type": "string"
            },
            "maxItems": 1,
            "type": "array"
          },
          "prompt": {
            "maxLength": 2000,
            "type": "string"
          },
          "roomName": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "callId",
          "roomName",
          "agentEpoch",
          "callerParticipantId",
          "agentParticipantId",
          "language",
          "permittedTools",
          "prompt"
        ],
        "type": "object"
      },
      "VoiceEventReceipt": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "additionalProperties": false,
        "properties": {
          "eventId": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "status": {
            "enum": [
              "committed",
              "duplicate"
            ],
            "type": "string"
          }
        },
        "required": [
          "eventId",
          "status"
        ],
        "type": "object"
      },
      "VoiceLease": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "additionalProperties": false,
        "properties": {
          "agentEpoch": {
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "type": "integer"
          },
          "callId": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
            "type": "string"
          },
          "token": {
            "maxLength": 256,
            "minLength": 32,
            "type": "string"
          }
        },
        "required": [
          "callId",
          "agentEpoch",
          "expiresAt",
          "token"
        ],
        "type": "object"
      },
      "VoiceLeaseClaim": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "additionalProperties": false,
        "properties": {
          "callId": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "dispatchId": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "participantId": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "roomName": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "callId",
          "roomName",
          "dispatchId",
          "participantId"
        ],
        "type": "object"
      },
      "VoiceLeaseRenew": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "additionalProperties": false,
        "properties": {
          "agentEpoch": {
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "type": "integer"
          }
        },
        "required": [
          "agentEpoch"
        ],
        "type": "object"
      },
      "VoiceTool": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "additionalProperties": false,
        "properties": {
          "agentEpoch": {
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "type": "integer"
          },
          "arguments": {},
          "invocationId": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "const": "create_business_request",
            "type": "string"
          }
        },
        "required": [
          "invocationId",
          "agentEpoch",
          "name",
          "arguments"
        ],
        "type": "object"
      },
      "WorkerEvent": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "oneOf": [
          {
            "additionalProperties": false,
            "properties": {
              "agentEpoch": {
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991,
                "type": "integer"
              },
              "callId": {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              "eventId": {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              "occurredAt": {
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
                "type": "string"
              },
              "payload": {
                "additionalProperties": false,
                "properties": {
                  "endMs": {
                    "anyOf": [
                      {
                        "maximum": 9007199254740991,
                        "minimum": 0,
                        "type": "integer"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "language": {
                    "maxLength": 32,
                    "minLength": 1,
                    "type": "string"
                  },
                  "participantId": {
                    "maxLength": 128,
                    "minLength": 1,
                    "type": "string"
                  },
                  "revision": {
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991,
                    "type": "integer"
                  },
                  "segmentId": {
                    "maxLength": 128,
                    "minLength": 1,
                    "type": "string"
                  },
                  "sequence": {
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991,
                    "type": "integer"
                  },
                  "speaker": {
                    "enum": [
                      "CALLER",
                      "AGENT",
                      "HUMAN"
                    ],
                    "type": "string"
                  },
                  "startMs": {
                    "anyOf": [
                      {
                        "maximum": 9007199254740991,
                        "minimum": 0,
                        "type": "integer"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "text": {
                    "maxLength": 4000,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "segmentId",
                  "revision",
                  "speaker",
                  "participantId",
                  "sequence",
                  "text",
                  "language"
                ],
                "type": "object"
              },
              "schemaVersion": {
                "const": 1,
                "type": "number"
              },
              "sourceSequence": {
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991,
                "type": "integer"
              },
              "type": {
                "const": "transcript.final",
                "type": "string"
              }
            },
            "required": [
              "schemaVersion",
              "eventId",
              "callId",
              "agentEpoch",
              "sourceSequence",
              "occurredAt",
              "type",
              "payload"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "agentEpoch": {
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991,
                "type": "integer"
              },
              "callId": {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              "eventId": {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              "occurredAt": {
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
                "type": "string"
              },
              "payload": {
                "additionalProperties": false,
                "properties": {
                  "participantId": {
                    "maxLength": 128,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "participantId"
                ],
                "type": "object"
              },
              "schemaVersion": {
                "const": 1,
                "type": "number"
              },
              "sourceSequence": {
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991,
                "type": "integer"
              },
              "type": {
                "const": "agent.ready",
                "type": "string"
              }
            },
            "required": [
              "schemaVersion",
              "eventId",
              "callId",
              "agentEpoch",
              "sourceSequence",
              "occurredAt",
              "type",
              "payload"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "agentEpoch": {
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991,
                "type": "integer"
              },
              "callId": {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              "eventId": {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              "occurredAt": {
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
                "type": "string"
              },
              "payload": {
                "additionalProperties": false,
                "properties": {
                  "code": {
                    "enum": [
                      "CONFIGURATION",
                      "TRANSIENT",
                      "FATAL",
                      "LEASE_LOST"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "code"
                ],
                "type": "object"
              },
              "schemaVersion": {
                "const": 1,
                "type": "number"
              },
              "sourceSequence": {
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991,
                "type": "integer"
              },
              "type": {
                "const": "agent.failed",
                "type": "string"
              }
            },
            "required": [
              "schemaVersion",
              "eventId",
              "callId",
              "agentEpoch",
              "sourceSequence",
              "occurredAt",
              "type",
              "payload"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "agentEpoch": {
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991,
                "type": "integer"
              },
              "callId": {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              "eventId": {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              "occurredAt": {
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
                "type": "string"
              },
              "payload": {
                "additionalProperties": false,
                "properties": {
                  "controlId": {
                    "maxLength": 128,
                    "minLength": 1,
                    "type": "string"
                  },
                  "state": {
                    "enum": [
                      "PAUSED",
                      "STOPPED"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "controlId",
                  "state"
                ],
                "type": "object"
              },
              "schemaVersion": {
                "const": 1,
                "type": "number"
              },
              "sourceSequence": {
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991,
                "type": "integer"
              },
              "type": {
                "const": "control.ack",
                "type": "string"
              }
            },
            "required": [
              "schemaVersion",
              "eventId",
              "callId",
              "agentEpoch",
              "sourceSequence",
              "occurredAt",
              "type",
              "payload"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "agentEpoch": {
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991,
                "type": "integer"
              },
              "callId": {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              "eventId": {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              "occurredAt": {
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
                "type": "string"
              },
              "payload": {
                "additionalProperties": false,
                "properties": {
                  "lastSourceSequence": {
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991,
                    "type": "integer"
                  }
                },
                "required": [
                  "lastSourceSequence"
                ],
                "type": "object"
              },
              "schemaVersion": {
                "const": 1,
                "type": "number"
              },
              "sourceSequence": {
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991,
                "type": "integer"
              },
              "type": {
                "const": "transcript.flushed",
                "type": "string"
              }
            },
            "required": [
              "schemaVersion",
              "eventId",
              "callId",
              "agentEpoch",
              "sourceSequence",
              "occurredAt",
              "type",
              "payload"
            ],
            "type": "object"
          }
        ]
      }
    },
    "securitySchemes": {
      "workerBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "MSVA voice worker contracts",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/calls/{callId}/context": {
      "get": {
        "security": [
          {
            "workerBearer": []
          }
        ]
      }
    },
    "/calls/{callId}/events": {
      "post": {
        "security": [
          {
            "workerBearer": []
          }
        ]
      }
    },
    "/calls/{callId}/lease/renew": {
      "post": {
        "security": [
          {
            "workerBearer": []
          }
        ]
      }
    },
    "/calls/{callId}/tools": {
      "post": {
        "security": [
          {
            "workerBearer": []
          }
        ]
      }
    },
    "/leases/claim": {
      "post": {
        "security": [
          {
            "workerBearer": []
          }
        ]
      }
    }
  }
}